refactor: introduce AbstractControllerConfiguration
