+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/config/AbstractControllerConfiguration.java
@@ -1,0 +1,80 @@
+package io.javaoperatorsdk.operator.api.config;
+
+import io.fabric8.kubernetes.client.CustomResource;
+import java.util.Collections;
+import java.util.Set;
+
+public abstract class AbstractControllerConfiguration<R extends CustomResource>
+    implements ControllerConfiguration<R> {
+
+  private final String associatedControllerClassName;
+  private final String name;
+  private final String crdName;
+  private final String finalizer;
+  private final boolean generationAware;
+  private final Set<String> namespaces;
+  private final boolean watchAllNamespaces;
+  private final RetryConfiguration retryConfiguration;
+
+  public AbstractControllerConfiguration(
+      String associatedControllerClassName,
+      String name,
+      String crdName,
+      String finalizer,
+      boolean generationAware,
+      Set<String> namespaces,
+      RetryConfiguration retryConfiguration) {
+    this.associatedControllerClassName = associatedControllerClassName;
+    this.name = name;
+    this.crdName = crdName;
+    this.finalizer = finalizer;
+    this.generationAware = generationAware;
+    this.namespaces =
+        namespaces != null ? Collections.unmodifiableSet(namespaces) : Collections.emptySet();
+    this.watchAllNamespaces = this.namespaces.contains(WATCH_ALL_NAMESPACES_MARKER);
+    this.retryConfiguration =
+        retryConfiguration == null
+            ? ControllerConfiguration.super.getRetryConfiguration()
+            : retryConfiguration;
+  }
+
+  @Override
+  public String getName() {
+    return name;
+  }
+
+  @Override
+  public String getCRDName() {
+    return crdName;
+  }
+
+  @Override
+  public String getFinalizer() {
+    return finalizer;
+  }
+
+  @Override
+  public boolean isGenerationAware() {
+    return generationAware;
+  }
+
+  @Override
+  public String getAssociatedControllerClassName() {
+    return associatedControllerClassName;
+  }
+
+  @Override
+  public Set<String> getNamespaces() {
+    return namespaces;
+  }
+
+  @Override
+  public boolean watchAllNamespaces() {
+    return watchAllNamespaces;
+  }
+
+  @Override
+  public RetryConfiguration getRetryConfiguration() {
+    return retryConfiguration;
+  }
+}
